+++ induct_shop/public/js/project.js
@@ -1,0 +1,5 @@
+frappe.ui.form.on('Project', {
+	refresh: function(frm) {
+		frm.set_df_property('monitor_progress_tab', 'hidden', 1);
+	}
+});